Add OASIS ga4 download stats client
oasis-download-stats-ga4-client.py
----------------------------------
Add python script to download OASIS download stats from GA4 property using service account credentials and GA4 property ID. The script takes the path to the credentials JSON file, GA4 property ID, and the number of UA downloads as recorded on UA before switching to GA4 as arguments. If the number of UA downloads is not provided, it defaults to 53948, which is the number of downloads from OASIS UA on 22nd Jun 2023. The script uses the Google Analytics Data API to run a simple report on the GA4 property and calculate the total number of downloads. The script writes the total number of downloads to a JSON file named oasis_download_stats.json.